// auto_api/models/flashcard-set.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Flashduo Api
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { Flashcard } from './flashcard';
import { FlashcardFolder } from './flashcard-folder';
import { User } from './user';
 /**
 * 
 *
 * @export
 * @interface FlashcardSet
 */
export interface FlashcardSet {

    /**
     * @type {number}
     * @memberof FlashcardSet
     */
    id?: number;

    /**
     * @type {string}
     * @memberof FlashcardSet
     */
    setName: string;

    /**
     * @type {string}
     * @memberof FlashcardSet
     */
    description?: string | null;

    /**
     * @type {boolean}
     * @memberof FlashcardSet
     */
    isPublic?: boolean;

    /**
     * @type {number}
     * @memberof FlashcardSet
     */
    flashcardFolderId?: number;

    /**
     * @type {FlashcardFolder}
     * @memberof FlashcardSet
     */
    flashcardFolder?: FlashcardFolder;

    /**
     * @type {number}
     * @memberof FlashcardSet
     */
    userId?: number;

    /**
     * @type {User}
     * @memberof FlashcardSet
     */
    user?: User;

    /**
     * @type {Date}
     * @memberof FlashcardSet
     */
    createdAt?: Date;

    /**
     * @type {Array<Flashcard>}
     * @memberof FlashcardSet
     */
    flashcards?: Array<Flashcard> | null;

    /**
     * @type {number}
     * @memberof FlashcardSet
     */
    totalCards: number;
}