// auto_api/models/flashcard.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Flashduo Api
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { FlashcardSet } from './flashcard-set';
import { User } from './user';
 /**
 * 
 *
 * @export
 * @interface Flashcard
 */
export interface Flashcard {

    /**
     * @type {number}
     * @memberof Flashcard
     */
    id?: number;

    /**
     * @type {string}
     * @memberof Flashcard
     */
    termLanguage: string;

    /**
     * @type {string}
     * @memberof Flashcard
     */
    definitionLanguage: string;

    /**
     * @type {string}
     * @memberof Flashcard
     */
    imageUrl?: string | null;

    /**
     * @type {string}
     * @memberof Flashcard
     */
    audioUrl?: string | null;

    /**
     * @type {boolean}
     * @memberof Flashcard
     */
    isFavorite?: boolean;

    /**
     * @type {boolean}
     * @memberof Flashcard
     */
    isDeleted?: boolean;

    /**
     * @type {number}
     * @memberof Flashcard
     */
    correctCount?: number;

    /**
     * @type {number}
     * @memberof Flashcard
     */
    wrongCount?: number;

    /**
     * @type {Date}
     * @memberof Flashcard
     */
    lastReviewedAt?: Date | null;

    /**
     * @type {number}
     * @memberof Flashcard
     */
    userId?: number;

    /**
     * @type {User}
     * @memberof Flashcard
     */
    user?: User;

    /**
     * @type {number}
     * @memberof Flashcard
     */
    flashcardSetId?: number;

    /**
     * @type {FlashcardSet}
     * @memberof Flashcard
     */
    flashcardSet?: FlashcardSet;

    /**
     * @type {Date}
     * @memberof Flashcard
     */
    createdAt?: Date;
}