// auto_api/models/flashcard-set-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Flashduo Api
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

 /**
 * 
 *
 * @export
 * @interface FlashcardSetRequest
 */
export interface FlashcardSetRequest {

    /**
     * @type {string}
     * @memberof FlashcardSetRequest
     */
    setName: string;

    /**
     * @type {string}
     * @memberof FlashcardSetRequest
     */
    description?: string | null;

    /**
     * @type {boolean}
     * @memberof FlashcardSetRequest
     */
    isPublic?: boolean;

    /**
     * @type {number}
     * @memberof FlashcardSetRequest
     */
    flashcardFolderId: number;

    /**
     * @type {number}
     * @memberof FlashcardSetRequest
     */
    userId: number;
}